app/requirements.txt
django==3.2
djangorestframework
pyTelegramBotAPI
psycopg2-binary
gunicorn
Celery
pyTelegramBotAPI